+++ src/helpers/utils/hash.ts
