+++ src/services/PineconeManager.ts
@@ -1,5 +1,5 @@
 import { Pinecone } from "@pinecone-database/pinecone";
-import obsidianFetchApi from "src/utils/obsidianFetchApi";
+import obsidianFetchApi from "src/helpers/utils/obsidianFetchApi";
 
 export const createPineconeClient = (apiKey: string) => {
 	return new Pinecone({
